Updating the optimized keword in file renaming and adjusting the imports
