+++ diffusers_helper/hunyuan_optimized.py
@@ -1,7 +1,7 @@
 import torch
 
 from diffusers.pipelines.hunyuan_video.pipeline_hunyuan_video import DEFAULT_PROMPT_TEMPLATE
-from diffusers_helper.utils import crop_or_pad_yield_mask
+from diffusers_helper.utils_optimized import crop_or_pad_yield_mask
 
 
 @torch.no_grad()
