+++ diffusers_helper/models/hunyuan_video_packed.py
@@ -15,7 +15,7 @@ from diffusers.models.embeddings import TimestepEmbedding, Timesteps, PixArtAlph
 from diffusers.models.modeling_outputs import Transformer2DModelOutput
 from diffusers.models.modeling_utils import ModelMixin
 from diffusers_helper.dit_common import LayerNorm
-from diffusers_helper.utils import zero_module
+from diffusers_helper.utils_optimized import zero_module
 
 
 enabled_backends = []
