+++ diffusers_helper/utils_optimized.py
